* android-4.19: (206154 commits)
  Linux 4.19.20
  cifs: Always resolve hostname before reconnecting
  md/raid5: fix 'out of memory' during raid cache recovery
  of: overlay: do not duplicate properties from overlay for new nodes
  of: overlay: use prop add changeset entry for property in new nodes
  of: overlay: add missing of_node_get() in __of_attach_node_sysfs
  of: overlay: add tests to validate kfrees from overlay removal
  of: Convert to using %pOFn instead of device_node.name
  mm: migrate: don't rely on __PageMovable() of newpage after unlocking it
  mm: hwpoison: use do_send_sig_info() instead of force_sig()
  mm, oom: fix use-after-free in oom_kill_process
  mm,memory_hotplug: fix scan_movable_pages() for gigantic hugepages
  oom, oom_reaper: do not enqueue same task twice
  mm/hugetlb.c: teach follow_hugetlb_page() to handle FOLL_NOWAIT
  kernel/exit.c: release ptraced tasks before zap_pid_ns_processes
  btrfs: On error always free subvol_name in btrfs_mount
  Btrfs: fix deadlock when allocating tree block during leaf/node split
  mmc: sdhci-iproc: handle mmc_of_parse() errors during probe
  platform/x86: asus-nb-wmi: Drop mapping of 0x33 and 0x34 scan codes
  platform/x86: asus-nb-wmi: Map 0x35 to KEY_SCREENLOCK
  ...

Conflicts:
	Documentation/devicetree/bindings/clock/rockchip,rk3399-cru.txt
	Documentation/devicetree/bindings/devfreq/event/rockchip-dfi.txt
	Documentation/devicetree/bindings/display/bridge/analogix_dp.txt
	Documentation/devicetree/bindings/display/rockchip/analogix_dp-rockchip.txt
	Documentation/devicetree/bindings/display/rockchip/dw_mipi_dsi_rockchip.txt
	Documentation/devicetree/bindings/display/rockchip/inno_hdmi-rockchip.txt
	Documentation/devicetree/bindings/display/rockchip/rockchip-lvds.txt
	Documentation/devicetree/bindings/media/i2c/ov2685.txt
	Documentation/devicetree/bindings/media/i2c/ov5695.txt
	Documentation/devicetree/bindings/media/i2c/ov7251.txt
	Documentation/devicetree/bindings/phy/phy-rockchip-inno-usb2.txt
	Documentation/devicetree/bindings/phy/phy-rockchip-typec.txt
	Documentation/devicetree/bindings/phy/rockchip-dp-phy.txt
	Documentation/devicetree/bindings/phy/rockchip-emmc-phy.txt
	Documentation/devicetree/bindings/soc/rockchip/grf.txt
	Documentation/devicetree/bindings/sound/rockchip,pdm.txt
	Documentation/devicetree/bindings/timer/rockchip,rk-timer.txt
	Documentation/devicetree/bindings/usb/rockchip,dwc3.txt
	arch/arm/boot/dts/rk3036-kylin.dts
	arch/arm/boot/dts/rk3036.dtsi
	arch/arm/boot/dts/rk3228-evb.dts
	arch/arm/boot/dts/rk3229-evb.dts
	arch/arm/boot/dts/rk322x.dtsi
	arch/arm/boot/dts/rk3288-fennec.dts
	arch/arm/boot/dts/rk3288-firefly-reload.dts
	arch/arm/boot/dts/rk3288-miqi.dts
	arch/arm/boot/dts/rk3288-phycore-rdk.dts
	arch/arm/boot/dts/rk3288-phycore-som.dtsi
	arch/arm/boot/dts/rv1108.dtsi
	arch/arm64/boot/dts/rockchip/rk3328-evb.dts
	arch/arm64/boot/dts/rockchip/rk3328-rock64.dts
	arch/arm64/boot/dts/rockchip/rk3328.dtsi
	arch/arm64/boot/dts/rockchip/rk3368-evb.dtsi
	arch/arm64/boot/dts/rockchip/rk3368-geekbox.dts
	arch/arm64/boot/dts/rockchip/rk3368-px5-evb.dts
	arch/arm64/boot/dts/rockchip/rk3399-gru.dtsi
	arch/arm64/boot/dts/rockchip/rk3399-opp.dtsi
	arch/arm64/boot/dts/rockchip/rk3399-sapphire.dtsi
	arch/arm64/boot/dts/rockchip/rk3399.dtsi
	drivers/clk/rockchip/clk-ddr.c
	drivers/clk/rockchip/clk-half-divider.c
	drivers/clk/rockchip/clk-px30.c
	drivers/clk/rockchip/clk-rk3036.c
	drivers/clk/rockchip/clk-rk3128.c
	drivers/clk/rockchip/clk-rk3228.c
	drivers/clk/rockchip/clk-rk3328.c
	drivers/clk/rockchip/clk-rk3399.c
	drivers/clk/rockchip/clk-rv1108.c
	drivers/devfreq/event/rockchip-dfi.c
	drivers/gpu/drm/bridge/analogix/Kconfig
	drivers/gpu/drm/bridge/analogix/Makefile
	drivers/gpu/drm/bridge/analogix/analogix_dp_core.c
	drivers/gpu/drm/bridge/analogix/analogix_dp_core.h
	drivers/gpu/drm/bridge/analogix/analogix_dp_reg.c
	drivers/gpu/drm/bridge/dumb-vga-dac.c
	drivers/gpu/drm/rockchip/analogix_dp-rockchip.c
	drivers/gpu/drm/rockchip/cdn-dp-core.c
	drivers/gpu/drm/rockchip/cdn-dp-core.h
	drivers/gpu/drm/rockchip/cdn-dp-reg.c
	drivers/gpu/drm/rockchip/cdn-dp-reg.h
	drivers/gpu/drm/rockchip/dw-mipi-dsi.c
	drivers/gpu/drm/rockchip/inno_hdmi.c
	drivers/gpu/drm/rockchip/inno_hdmi.h
	drivers/gpu/drm/rockchip/rockchip_lvds.c
	drivers/gpu/drm/rockchip/rockchip_vop_reg.c
	drivers/gpu/drm/rockchip/rockchip_vop_reg.h
	drivers/hid/hid-alps.c
	drivers/iio/light/vl6180.c
	drivers/leds/leds-is31fl32xx.c
	drivers/media/cec/cec-adap.c
	drivers/media/cec/cec-api.c
	drivers/media/cec/cec-notifier.c
	drivers/media/i2c/ov5647.c
	drivers/media/i2c/ov5695.c
	drivers/media/i2c/ov7251.c
	drivers/media/platform/rockchip/rga/rga.c
	drivers/media/rc/ir-imon-decoder.c
	drivers/media/rc/serial_ir.c
	drivers/media/spi/Kconfig
	drivers/media/spi/Makefile
	drivers/media/v4l2-core/v4l2-fwnode.c
	drivers/net/phy/rockchip.c
	drivers/phy/rockchip/Kconfig
	drivers/phy/rockchip/Makefile
	drivers/phy/rockchip/phy-rockchip-emmc.c
	drivers/phy/rockchip/phy-rockchip-inno-usb2.c
	drivers/phy/rockchip/phy-rockchip-typec.c
	drivers/phy/rockchip/phy-rockchip-usb.c
	drivers/pinctrl/pinctrl-rk805.c
	drivers/power/reset/reboot-mode.c
	drivers/soc/rockchip/grf.c
	drivers/usb/dwc3/dwc3-of-simple.c
	drivers/usb/gadget/udc/core.c
	include/drm/bridge/analogix_dp.h
	include/dt-bindings/clock/px30-cru.h
	include/dt-bindings/clock/rk3036-cru.h
	include/dt-bindings/clock/rk3128-cru.h
	include/dt-bindings/clock/rk3228-cru.h
	include/dt-bindings/clock/rk3328-cru.h
	include/dt-bindings/clock/rk3399-cru.h
	include/dt-bindings/power/px30-power.h
	include/dt-bindings/power/rk3036-power.h
	include/dt-bindings/power/rk3228-power.h
	include/media/cec-notifier.h
	include/soc/rockchip/rockchip_sip.h
	include/sound/hdmi-codec.h
	sound/soc/codecs/hdmi-codec.c
	sound/soc/rockchip/rockchip_pdm.c
	sound/soc/rockchip/rockchip_pdm.h

+++ tools/testing/selftests/efivarfs/efivarfs.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
+# SPDX-License-Identifier: GPL-2.0
 
 efivarfs_mount=/sys/firmware/efi/efivars
 test_guid=210be57c-9849-4fc7-a635-e6382d1aec27
+
+# Kselftest framework requirement - SKIP code is 4.
+ksft_skip=4
 
 check_prereqs()
 {
@@ -9,12 +13,12 @@ check_prereqs()
 
 	if [ $UID != 0 ]; then
 		echo $msg must be run as root >&2
-		exit 0
+		exit $ksft_skip
 	fi
 
 	if ! grep -q "^\S\+ $efivarfs_mount efivarfs" /proc/mounts; then
 		echo $msg efivarfs is not mounted on $efivarfs_mount >&2
-		exit 0
+		exit $ksft_skip
 	fi
 }
 
@@ -88,7 +92,11 @@ test_delete()
 		exit 1
 	fi
 
-	rm $file
+	rm $file 2>/dev/null
+	if [ $? -ne 0 ]; then
+		chattr -i $file
+		rm $file
+	fi
 
 	if [ -e $file ]; then
 		echo "$file couldn't be deleted" >&2
@@ -111,6 +119,7 @@ test_zero_size_delete()
 		exit 1
 	fi
 
+	chattr -i $file
 	printf "$attrs" > $file
 
 	if [ -e $file ]; then
@@ -141,7 +150,11 @@ test_valid_filenames()
 			echo "$file could not be created" >&2
 			ret=1
 		else
-			rm $file
+			rm $file 2>/dev/null
+			if [ $? -ne 0 ]; then
+				chattr -i $file
+				rm $file
+			fi
 		fi
 	done
 
@@ -174,7 +187,11 @@ test_invalid_filenames()
 
 		if [ -e $file ]; then
 			echo "Creating $file should have failed" >&2
-			rm $file
+			rm $file 2>/dev/null
+			if [ $? -ne 0 ]; then
+				chattr -i $file
+				rm $file
+			fi
 			ret=1
 		fi
 	done
